requirements.txt
docker==5.0.0
python-powerdns==0.2.1
